requirements.txt
# Core automation libraries
pyautogui==0.9.54
pywinauto==0.6.8
keyboard==0.13.5

# Image processing (NumPy 버전 고정 - opencv 호환성)
numpy<2.0.0
opencv-python==4.8.1.78
Pillow==10.1.0

# Window management
pywin32==306
psutil==5.9.8

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pyperclip==1.8.2

# OCR for monitoring processing status (optional, not used with icon detection)
# easyocr==1.7.1
# torch==2.0.1
# torchvision==0.15.2

